Change rendering keymap context in helpscreen
I agree with Huevos that this is indeed kind of tricky and after the
release of 9.1 indeed some strings were found that were not correctly
transformed from camalcase to a good readable string with spaces. E.g.
when CCCam is splitted and we had more examples. This at least better
works then what I initially did suggested.